Move codeflow test mocks to a namespace (#5324)
