+++ test/ProductConstructionService.DependencyFlow.Tests/Mocks/MockRedisCacheFactory.cs
@@ -4,7 +4,7 @@
 using Moq;
 using ProductConstructionService.Common;
 
-namespace ProductConstructionService.DependencyFlow.Tests;
+namespace ProductConstructionService.DependencyFlow.Tests.Mocks;
 
 internal class MockRedisCacheFactory : IRedisCacheFactory
 {
