+++ test/ProductConstructionService.DependencyFlow.Tests/UpdaterTests.cs
@@ -13,6 +13,7 @@ using Moq;
 using NUnit.Framework;
 using ProductConstructionService.Common;
 using ProductConstructionService.DependencyFlow.Model;
+using ProductConstructionService.DependencyFlow.Tests.Mocks;
 using ProductConstructionService.WorkItems;
 
 namespace ProductConstructionService.DependencyFlow.Tests;
